requirements.txt
playwright==1.40.0
python-telegram-bot==20.8
nest_asyncio==1.6.0
pymupdf==1.23.22

